wasm backend: Use -flto rather then -emit-llvm when generating bitcode objects
With the wasm backend we use lld to combine bitcode objects and
using -flto rather then -emit-llvm provides the linker with more
information. Specifically it provides thin LTO information that
that linker can use as needed.

Also, only pass -lto-O linker flag to disable cross-module-optimization
when we are not optimizing.  Otherwise let the linker default to
lto-O2.

In general the LTO compiler will optimize according the bitcode inputs
that its given, not this flag. (i.e. it should honer the "optnone"
bitcode attribute which clang generates at -O0).